prevouts: add the sequence to RevaultPrevouts, rename them to TxIns
This was coming. PrevOuts can be reused and with the inclusion of the
TxOut into them we lost the Copy, intuitively bounding their lifetimes
to one usage.

Besides, we had those monstruous inputs in RevaultTransactions'
new()s.

Now that their sequence is bounded to them, they are one-shot
transaction inputs, and their usage is much more intuitive!

Of course, this helps me with PSBT as well :-)

Signed-off-by: Antoine Poinsot <darosior@protonmail.com>