txins: only take the sequence in new() for the unvault txin
Letting them specify it for the other txins is just YA source of useless
confusion.

I just used 0xff_ff_ff_fd for all of them, as i far as i can imagine
this is not going to cause any issue for us (and if it did, it'd be
quite worrying tbh).

Signed-off-by: Antoine Poinsot <darosior@protonmail.com>