scripts: generalize the unvault descriptor policy
This changes the API to provide keys for the *stakeholders* and not
anymore the non-fund-managers only.

Therefore we segregate the set of keys of the stakeholders and the one
of the fund managers, breaking the optimizations allowed by the
assumption that the latter was actually a *subset* of the former.

Further, this allows to:
- have external fund managers (eg employees).
- have a treshold of fund managers.

Signed-off-by: Antoine Poinsot <darosior@protonmail.com>