Upgrade to latest Miniscript
This is the first part of a big refactoring, considering recent
advancement and API breaks in the Miniscript library.
It unfortunately is a huge commit to keep bisectability.

We first get rid of our boutique satisfier to use The Official :tm: one.

We adjust our functions to the new API that take a Secp256k1 context for
derivation, and an "xpub context" (basically the secp256k1 context +
child number if wildcard) for -well- xpubs.

We also get rid of the bitcoin dependency to be in sync with
Miniscript's.

We get rid of the now-useless raw_unvault_descriptor() function.

We restablish (FINALLY !) the bounds checks now that the limits were
implemented and fixed upstream.

Fixes #29
Fixes #25

Signed-off-by: Antoine Poinsot <darosior@protonmail.com>