transactions: external transactions to not store a Psbt anymore
Storing a Psbt for a transaction that we not going to sign is
nonsensical.

In addition, VaultTransaction and FeeBumpTransaction have always been
special citizens: we don't sign them and are external (VaultTransaction
from the payer and FeeBumpTransaction from the managers' wallet).
This makes it official by removing the implementation of
RevaultTransaction for them. We just use a thin newtype for them.

Signed-off-by: Antoine Poinsot <darosior@protonmail.com>