transactions: split signature_hash in internal/external versions
Internally, we only ever manage P2WSH and always set the witness script
for PSBT inputs. But we still take the scriptCode as input in case we
are sighash'ing an external (feebump) input.

So, split the method in one script_code-requiring version, and one that
can safely unwrap().

Also, this makes the methods return an Option: caller is smart enough
to not need a detailed String message for a missing input..

Signed-off-by: Antoine Poinsot <darosior@protonmail.com>