transactions: assert on sane state rather than returning None/Err
If we don't do it, who's going to ?

We know for sure (well we don't until parsing checks but hey, coming)
that we can't sanely create such a transaction, so we are the most
informed to assert.

They could in theory tweak the PSBT with inner_tx_mut but... That'd be
a bad idea (note to self: remove inner_tx_mut()).

Signed-off-by: Antoine Poinsot <darosior@protonmail.com>