Tree-wide: update rust-miniscript to the latest major version
This release of rust-miniscript featured an entire refactor of the type
system, which lead to pretty invasive changes here.

The main modification of the API is the introduction of a new newtype
for each descriptor ('concrete' ones, aka already-derived) and
restricting the existing types to only contain wildcard-xpubs.

This further allowed to contain the derivation in the descriptor types
themselves, therefore removing any secp context argument of the
transactions creation routines.

Signed-off-by: Antoine Poinsot <darosior@protonmail.com>