txouts: make FeeBumpTxOut a newtype
Re-using the macro and storing a witness_script that is always going to
be None is just bad design.

By trying to get rid of the Optional witness_script i also thought about
abusing it for the FeeBumpTxOut by storing the scriptCode (a P2PKH,
actually) but it's confusing.. The point is that they are fundamentally
different txouts and i probably drew us in a corner here by forcing both
types to be under the same trait.

Signed-off-by: Antoine Poinsot <darosior@protonmail.com>