txouts: a trait to differentiate internal txos (we know the witness Script)
This also gets rid of the create_tx! macro in order to create PsbtIns
without witness_script for the feebump inputs. This is arguably another
win, as even though it made the code more concise it was confusing.

We also get rid of the witness_script in the PSBT outputs: we don't use
them in practice (we keep the descriptor and derivation index), so no
need to waste space.

Signed-off-by: Antoine Poinsot <darosior@protonmail.com>