script: add key source info to keys in derived descriptors
We'll need this information to be stored in the PSBT inputs of the
transactions, to eventually be passed to HWs.

This slightly changes the str representation to not loose the source
information when parsing. Of course it means we won't be able to parse a
derived descriptor directly from the chain with no more information but
it's arguably not very useful.

Signed-off-by: Antoine Poinsot <darosior@protonmail.com>