transactions: improve the CPFP output detection in CpfpableTransaction
We might not know in advance the derivation index of the CPFP output.
Since we now store the bip32 derivations in the PSBT outs, we can take
advantage of this in order to never miss a CPFP output that is present
in the transaction, at the expense of more computation.

More computation would only realistically happen for the Spend -since
the Unvault only has 2 outputs-, but in this case only 2 outputs
effectively have the bip32 derivations set, so we would only derive the
descriptor twice.