transactions: introduce a RevaultPresignedTransaction trait
There is conceptually a distinction between the Unvault, Cancel,
Emergency's on one hand and the Spend on the other hand. We treat them
differently in revaultd, and we've already been in the situation where
we wanted a function to generalistically take a Cancel, Unvault,
Emergency but not a Spend transaction.

However, it was previously convoluted to introduce a type distinction
between those, as the Unvault was still pretty different from the
revocation transactions.

Now that all the presigned transactions are single-input and signed with
ALL, this is possible to merge the common implementation of sighash() and
add_sig() for the first and single input into a single trait. This also
enables the type distinction that will turn out to be useful downstream.